More clean up of Visual Studio project / solution files.
git-svn-id: http://svn.osgeo.org/mapguide/trunk@1884 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/MapGuideCommon/MapLayer/Layer.cpp
@@ -308,8 +308,7 @@ MgPropertyCollection* MgLayer::UpdateFeatures(MgFeatureCommandCollection* comman
 
     // Create the feature source object.
     MdfParser::FSDSAX2Parser parser;
-    parser.ParseString(xmlContent.c_str(),
-        (unsigned int)(xmlContent.length() * sizeof(char)));
+    parser.ParseString(xmlContent.c_str(), xmlContent.length() * sizeof(char));
     ACE_ASSERT(parser.GetSucceeded());
 
     auto_ptr<MdfModel::FeatureSource> featureSource;
